removed dependencies BookDao & ReaderDao from LibraryDao and added them to LibraryService
