+++ src/main/java/company/name/dao/ReaderDao.java
@@ -1,5 +1,6 @@
 package company.name.dao;
 
+import company.name.models.Book;
 import company.name.models.Reader;
 import java.util.List;
 
@@ -11,4 +12,9 @@ public interface ReaderDao {
     List<Reader> getAll();
 
     void update(Reader reader);
+
+    List<Long> getBorrowedBooksIds(Reader reader);
+
+    Reader getCurrentReaderOfBook(Book book);
+
 }
